requirements.txt
openai==1.55.3
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.36.0
